// frontend/src/api/schema.gen.ts
export interface components {
    schemas: {
// ...
        ShiftGroupOut: {
            /** Code */
            code: string;
            /**
             * Is Work Shift
             * @default true
             */
            is_work_shift: boolean;
            /**
             * Is Night Shift
             * @default false
             */
            is_night_shift: boolean;
            /** Id */
            id: number;
        };
        /** ShiftGroupUpdate */
        ShiftGroupUpdate: {
            /** Code */
            code?: string | null;
            /** Is Work Shift */
            is_work_shift?: boolean | null;
            /** Is Night Shift */
            is_night_shift?: boolean | null;
        };
        /** ShiftOut */
        ShiftOut: {
            /** Group Id */
            group_id: number;
            /** Code */
            code: string;
            /** Name */
            name: string;
            /** Start Min */
            start_min: number;
            /** End Min */
            end_min: number;
            /** Work Min */
            work_min: number;
            /**
             * Break Min
             * @default 0
             */
            break_min: number;
            /** Id */
            id: number;
            group: components["schemas"]["ShiftGroupOut"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}